Expand functionality to read both passed and failed tests
qUnit uses ".pass" and ".fail" class to mark whether a module or a test
passed or failed. I changed the methods to take a parameter that allows
the class to be specify whether we want to read passed or failed tests.
The .init() method then call the _reResults twice, once with ".pass"
and again with ".fail" to get both results.
A module will fail even if only one assertion fails. This script,
however, will read for pass tests in a failed module as well. That
means a single module may appear twice if a module contains both passed
and failed tests.