Change the returned object format
Change the format of the returned JSON object to use static property
names.
It turns out Selenium converts all returned value automatically using
json_encode( ). However, it does so to an object instead of array.
Working with objects that have dynamic name with a space in it like
"module 1" is difficult.
The object now has property names like "name" and "number" instead.