Combine tests from the same module together
The returned object is not a little more conscious about the module. If
the next test has the same module name as the previous one, it will add
the tests to the previous module in the returned object.
This is still not perfect though. If you have a module that exists in 2
non-consecutive fashion like:
module A
module B
module A
, this script will not be able to combine module A *yet*.