Change var name from module to test
In _readResults, we're really working with each test shown to user. The
concept of module hierarchy only exists in the source code but not
display clearly. Renaming variable to tests would be best. As mentioned
in previous commit, if we want info about module, it will need to be
computed. Will create new methods for handling those in the future.